Refacto to isaolate css calculations in ODSChartsCssHelper
