+++ src/theme/ods-chart-theme.ts
@@ -11,7 +11,7 @@ import { COMMON_LINE_STYLE_POINTS } from './common/ODS.line-style.with-points';
 import { COMMON_LINE_STYLE_SMOOTH } from './common/ODS.line-style.smooth';
 import { EChartsProject, ODS_PROJECT } from './ODS.project';
 import { ODSChartsLegends } from './legends/ods-chart-legends';
-import { isVarArray, isVarObject, mergeObjects } from '../tools/merge-objects';
+import { mergeObjects } from '../tools/merge-objects';
 import { ODSChartsResize } from './resize/ods-chart-resize';
 import { ODSChartsCSSThemeDefinition, ODSChartsCSSThemes, ODSChartsCSSThemesNames } from './css-themes/css-themes';
 import { getStringValue } from '../tools/hash';
@@ -20,6 +20,9 @@ import { ODSChartsPopover } from './popover/ods-chart-popover';
 import { ODSChartsPopoverConfig, ODSChartsPopoverDefinition, ODSChartsPopoverManagers } from './popover/ods-chart-popover-definitions';
 import { isMainAxis } from './const/main-axis-type.const';
 import { ODSChartsLegendHolderDefinition } from './legends/ods-chart-legends-definitions';
+import { ODSChartsThemeObserver } from './theme-observer/ods-charts-theme-observer';
+import { ODSChartsCssHelper } from './css-themes/css-helper';
+
 import { DEFAULT_COLORS_YELLOW } from './default/ODS.colors.yellow';
 import { DEFAULT_COLORS_PURPLE } from './default/ODS.colors.purple';
 import { DEFAULT_COLORS_PINK } from './default/ODS.colors.pink';
@@ -33,9 +36,6 @@ import { DEFAULT_COLORS_CATEGORICAL } from './default/ODS.colors.categorical';
 import { DEFAULT_COLORS } from './default/ODS.colors';
 import { DEFAULT_COMMON } from './default/ODS.common';
 import { DEFAULT_LINES_AXIS } from './default/ODS.lines.axis';
-import { ODS_CHARTS_CSS_VARIABLES } from './colors/colors-css-variables';
-import { ODSChartsThemeObserver } from './theme-observer/ods-charts-theme-observer';
-
 import { DEFAULT_OUDS_COLORS_BLUE } from './default/OUDS.colors.blue';
 import { DEFAULT_OUDS_COLORS_CATEGORICAL } from './default/OUDS.colors.categorical';
 import { DEFAULT_OUDS_COLORS_FUNCTIONAL } from './default/OUDS.colors.functional';
@@ -45,7 +45,6 @@ import { DEFAULT_OUDS_COLORS_PINK } from './default/OUDS.colors.pink';
 import { DEFAULT_OUDS_COLORS_PURPLE } from './default/OUDS.colors.purple';
 import { DEFAULT_OUDS_COLORS_SINGLE } from './default/OUDS.colors.single';
 import { DEFAULT_OUDS_COLORS_YELLOW } from './default/OUDS.colors.yellow';
-import { conditionalCloneDeepObject } from '../tools/conditional-clone-deep-object';
 // import { DEFAULT_OUDS_COMMON } from './default/OUDS.common'; // TODO: use when we can switch between ODS and OUDS
 // import { DEFAULT_OUDS_LINES_AXIS } from './default/OUDS.lines.axis';
 
@@ -77,6 +76,7 @@ import { conditionalCloneDeepObject } from '../tools/conditional-clone-deep-obje
  * - {@link OUDS_HIGHLIGHT} is the color set embedding all the Orange Unified Design System highlight colors. It's still experimental.
  * - {@link OUDS_SINGLE} is the color set embedding all the Orange Unified Design System single colors. It's still experimental.
  */
+
 export enum ODSChartsColorsSet {
   DEFAULT = 'default',
   CATEGORICAL = 'categorical',
@@ -306,11 +306,8 @@ export class ODSChartsTheme {
   private chartPopoverManager: ODSChartsPopover = undefined as unknown as ODSChartsPopover;
   private chartThemeObserver: ODSChartsThemeObserver = undefined as unknown as ODSChartsThemeObserver;
   public cssThemeName: ODSChartsCSSThemesNames;
-  private cssVarsMapping: { [variable: string]: string } = {};
   public theme: EChartsProject;
-
-  private _computedStyle: CSSStyleDeclaration | undefined | null = undefined;
-  private _computedStyleInitialized: boolean = false;
+  private cssHelper: ODSChartsCssHelper;
 
   private constructor(
     public name: string,
@@ -321,202 +318,9 @@ export class ODSChartsTheme {
       (Object.keys(ODSChartsCSSThemes).find(
         (oneTheme) => ODSChartsCSSThemes[oneTheme as ODSChartsCSSThemesNames] === options.cssTheme
       ) as ODSChartsCSSThemesNames) || ODSChartsCSSThemesNames.CUSTOM;
+
+    this.cssHelper = new ODSChartsCssHelper(this.options.cssSelector as string, this.options.mode, this.cssThemeName);
     this.theme = this.calculateTheme();
-  }
-
-  /**
-   * build a CSSStyleDeclaration build from a DOM element insert in the current charts holder.
-   * It is used to calculate real time values for the css var.
-   */
-  private get computedStyle(): CSSStyleDeclaration | undefined {
-    if (undefined === this._computedStyle) {
-      if (this.options.cssSelector) {
-        let contextElement = document.querySelector(
-          this.options.cssSelector + '>.ods-charts-style-' + (this.options.mode ? this.options.mode : ODSChartsMode.DEFAULT)
-        );
-        if (!contextElement) {
-          const domElement: Element = (this.options.cssSelector ? document.querySelector(this.options.cssSelector) : null) as Element;
-          if (domElement) {
-            contextElement = document.createElement('div');
-            if (this.options.mode && this.options.mode !== ODSChartsMode.DEFAULT) {
-              contextElement.setAttribute('data-bs-theme', this.options.mode);
-            }
-            contextElement.classList.add('ods-charts-context');
-            contextElement.classList.add('ods-charts-style-' + (this.options.mode ? this.options.mode : ODSChartsMode.DEFAULT));
-            domElement.append(contextElement);
-          }
-        }
-        if (contextElement) {
-          this._computedStyle = window.getComputedStyle(contextElement);
-        }
-      }
-      this._computedStyle = this._computedStyle ? this._computedStyle : null;
-      if (!this._computedStyle) {
-        console.error('unable to build computed style for chart css context', this.options.cssSelector, this.options.cssSelector);
-      }
-    }
-    return this._computedStyle ? this._computedStyle : undefined;
-  }
-
-  /**
-   * Init DOM elemnts for style. Add any specific style for the current theme.
-   * Register a computed style will be initialised.
-   * if `initComputedStyle()` return true, `removeComputedStyle()` must be called`after current treatment is done.
-   * @returns
-   */
-  private initComputedStyle(): boolean {
-    if (!this._computedStyleInitialized) {
-      if (!document.getElementById('ods-charts-style-' + this.cssThemeName) && ODS_CHARTS_CSS_VARIABLES[this.cssThemeName]) {
-        const style = document.createElement('style');
-        style.textContent = ODS_CHARTS_CSS_VARIABLES[this.cssThemeName];
-        style.id = 'ods-charts-style-' + this.cssThemeName;
-        document.querySelector('head')?.append(style);
-      }
-      this._computedStyleInitialized = true;
-      return true;
-    }
-    return false;
-  }
-
-  /**
-   * After replacing all the css var of used objects by their value calculated in the DOM,
-   * remove the used temporay DOM element.
-   * `removeComputedStyle()` must be called if and only if the previous `initComputedStyle()` call returns true.
-   */
-  private removeComputedStyle() {
-    if (this.options.cssSelector) {
-      const contextElement = document.querySelector(
-        this.options.cssSelector + '>.ods-charts-style-' + (this.options.mode ? this.options.mode : ODSChartsMode.DEFAULT)
-      );
-      if (contextElement) {
-        contextElement.remove();
-      }
-    }
-    this._computedStyle = undefined;
-    this._computedStyleInitialized = false;
-  }
-
-  /**
-   * give the calculated value of a property
-   * @param name : the property name
-   * @returns the value of that property in the style computed from the DOM
-   */
-  private getPropertyValue(name: string): string {
-    const value = this.computedStyle ? this.computedStyle.getPropertyValue(name) : '';
-    return value ? value : '';
-  }
-
-  /**
-   * Replace a css var by it value
-   * @param css the string to be analysed
-   * @returns the updated value
-   */
-  private replaceOneCssVar(css: any) {
-    let returnedValue = css;
-    if (this.options.cssSelector && 'string' === typeof returnedValue && !!this.computedStyle) {
-      try {
-        const matches = this.getCssVarMatch(returnedValue);
-        if (matches) {
-          for (const foundVar of matches) {
-            if (!(foundVar in this.cssVarsMapping)) {
-              const varPartsRex = /var\( ?([^, ]+) ?, ?([^ ]+) ?\)/;
-              const varParts = foundVar.match(varPartsRex);
-              if (varParts) {
-                const varValue = this.getPropertyValue(varParts[1]);
-                if (!varValue) {
-                  console.error('missing css var ', foundVar);
-                }
-                this.cssVarsMapping[foundVar] = varValue ? varValue : varParts[2];
-              } else {
-                const varNameRex = /var\( ?(.+) ?\)/;
-                const varName = foundVar.match(varNameRex);
-                if (varName) {
-                  const varValue = this.getPropertyValue(varName[1]);
-                  if (varValue) {
-                    this.cssVarsMapping[foundVar] = varValue;
-                  } else {
-                    console.error('missing css var ', foundVar);
-                    this.cssVarsMapping[foundVar] = foundVar;
-                  }
-                } else {
-                  console.error('missing css var ', foundVar);
-                  this.cssVarsMapping[foundVar] = foundVar;
-                }
-              }
-            }
-
-            if (foundVar in this.cssVarsMapping) {
-              returnedValue = returnedValue.replace(foundVar, this.cssVarsMapping[foundVar]);
-            }
-          }
-        }
-      } catch (error) {}
-    }
-    return returnedValue;
-  }
-
-  /**
-   * replace any css var reference by its value
-   * @param subTreeConfig: subTree in which css var must be replaced by their value
-   * @returns the subTreeConfig updated
-   */
-  private replaceRecursivelyCssVars<T extends { [key: string]: any }>(subTreeConfig: T): T {
-    var newConfig: { [key: string]: any } = subTreeConfig;
-    for (const key of Object.keys(newConfig)) {
-      if (isVarArray(newConfig[key])) {
-        for (let index = 0; index < newConfig[key].length; index++) {
-          if (isVarObject(newConfig[key][index])) {
-            newConfig[key][index] = this.replaceRecursivelyCssVars(newConfig[key][index]);
-          } else if ('string' === typeof newConfig[key][index]) {
-            newConfig[key][index] = this.replaceOneCssVar(newConfig[key][index]);
-          }
-        }
-      } else if (isVarObject(newConfig[key])) {
-        newConfig[key] = this.replaceRecursivelyCssVars(newConfig[key]);
-      } else if ('string' === typeof newConfig[key]) {
-        newConfig[key] = this.replaceOneCssVar(newConfig[key]);
-      }
-    }
-    return newConfig as T;
-  }
-
-  /**
-   * Return matches of a css var in the given string
-   * @param value : string in whic looking for in css var
-   * @returns : the regular expresion matches
-   */
-  private getCssVarMatch(value: string): RegExpMatchArray | null {
-    const regex = /var\(([^,]*),?(.*)\)/g;
-    return value.match(regex);
-  }
-
-  /**
-   * Replace any css var reference (like `var(--bs-border-color-subtle)`) by its value.
-   * @param themeConfiguration : object containing some css var string refernce
-   * @returns : the themeCongiguration updated
-   */
-  private replaceAllCssVars<T extends { [key: string]: any }>(themeConfiguration: T): T {
-    let result = themeConfiguration;
-
-    const initStyleNotYetDone = this.initComputedStyle();
-    if (!!this.computedStyle) {
-      result = this.replaceRecursivelyCssVars(themeConfiguration);
-    }
-    if (initStyleNotYetDone) {
-      this.removeComputedStyle();
-    }
-
-    return result;
-  }
-
-  /**
-   * Make a copy of any object parts that contains css var reference (like `var(--bs-border-color-subtle)`)
-   * and replace those css vars by their value.
-   * @param chartData : object containing some css var string reference
-   * @returns the new object partially clone
-   */
-  private cloneAndReplaceAllCssVars<T extends { [key: string]: any }>(chartData: T): T {
-    return this.replaceAllCssVars(conditionalCloneDeepObject(chartData, (value) => !!this.getCssVarMatch(value)));
   }
 
   /**
@@ -525,8 +329,7 @@ export class ODSChartsTheme {
    * @returns the updated current theme
    */
   private calculateTheme(): EChartsProject {
-    this.cssVarsMapping = {};
-    this.theme = this.replaceAllCssVars(cloneDeepObject(this.initialTheme));
+    this.theme = this.cssHelper.replaceAllCssVars(cloneDeepObject(this.initialTheme));
     return this.theme;
   }
 
@@ -786,15 +589,16 @@ export class ODSChartsTheme {
 
     if (this.chartThemeObserver) {
       this.options.mode = ODSChartsTheme.getDarkOrLightMode(this.chartThemeObserver.addThemeObserver());
+      this.cssHelper.changeThemeMode(this.options.mode);
     }
 
-    const hasInitializedCompitedStyle = this.initComputedStyle();
+    const hasInitializedCompitedStyle = this.cssHelper.initComputedStyle();
     try {
       // need to copy dataOptions as theme feature may change the original dataOptions
       // only make an asign at the first level in order to avoid deep copy of all data
       // each feature is responsible to deep copy the changed part according to their changes.
       // Also make a deep copy when needed to replace any css var
-      const updatedDataOptionsForTheme = this.cloneAndReplaceAllCssVars({ ...this.dataOptions });
+      const updatedDataOptionsForTheme = this.cssHelper.cloneAndReplaceAllCssVars({ ...this.dataOptions });
 
       const axisLabel = {
         fontStyle: 'normal',
@@ -877,7 +681,7 @@ export class ODSChartsTheme {
 
       const displayedColors = this.getDisplayedColors(usedTheme.color, updatedDataOptionsForTheme);
 
-      themeOptions = this.replaceAllCssVars(themeOptions);
+      themeOptions = this.cssHelper.replaceAllCssVars(themeOptions);
 
       if (this.chartLegendManager) {
         try {
@@ -918,7 +722,7 @@ export class ODSChartsTheme {
       return { themeOptions, dataOptions: updatedDataOptionsForTheme };
     } finally {
       if (hasInitializedCompitedStyle) {
-        this.removeComputedStyle();
+        this.cssHelper.removeComputedStyle();
       }
     }
   }
